+++ pom.xml
@@ -98,7 +98,7 @@
         <module>simulator</module>
         <module>tests</module>
         <module>utils</module>
-        <module>visualiser</module>
+        <module>visualizer</module>
     </modules>
 
     <scm>
