+++ visualizer/src/main/java/com/hazelcast/simulator/visualizer/Main.java
@@ -13,24 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hazelcast.simulator.visualiser.ui;
+package com.hazelcast.simulator.visualizer;
 
-import com.hazelcast.simulator.visualiser.data.Model;
+import com.hazelcast.simulator.visualizer.ui.MainUI;
 
 import javax.swing.*;
-import java.awt.*;
 
-public class LoadedBenchmarks extends JPanel implements Model.BenchmarkChangeListener {
+public final class Main {
 
-    public LoadedBenchmarks() {
-        setMaximumSize(new Dimension(Integer.MAX_VALUE, Integer.MAX_VALUE));
-        setBorder(BorderFactory.createTitledBorder("Benchmarks"));
-        setLayout(new BoxLayout(this, BoxLayout.Y_AXIS));
+    private Main() {
     }
 
-    @Override
-    public void benchmarkChanged(String benchmarkName) {
-        JLabel label = new JLabel(benchmarkName);
-        add(label);
+    public static void main(String[] args) {
+        SwingUtilities.invokeLater(new Runnable() {
+            public void run() {
+                new MainUI();
+            }
+        });
     }
 }
