+++ visualizer/src/main/java/com/hazelcast/simulator/visualizer/data/SimpleHistogramDataSetContainer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hazelcast.simulator.visualiser.data;
+package com.hazelcast.simulator.visualizer.data;
 
 import org.jfree.data.statistics.SimpleHistogramBin;
 import org.jfree.data.statistics.SimpleHistogramDataset;
