+++ visualizer/src/main/java/com/hazelcast/simulator/visualizer/ui/MainUI.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hazelcast.simulator.visualiser.ui;
+package com.hazelcast.simulator.visualizer.ui;
 
-import com.hazelcast.simulator.visualiser.data.Model;
-import com.hazelcast.simulator.visualiser.io.ResultParserWorker;
+import com.hazelcast.simulator.visualizer.data.Model;
+import com.hazelcast.simulator.visualizer.io.ResultParserWorker;
 import org.jfree.ui.ExtensionFileFilter;
 
 import javax.swing.*;
