+++ visualizer/src/main/java/com/hazelcast/simulator/visualizer/utils/ListenerUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hazelcast.simulator.visualiser.utils;
+package com.hazelcast.simulator.visualizer.utils;
 
 import javax.swing.*;
 import java.awt.*;
